+++ src/pipe.test.js
@@ -1,11 +1,11 @@
-import compose from './compose'
+import pipe from './pipe'
 
-describe('compose', () => {
+describe('pipe', () => {
   it('composes the given functions', () => {
     function f (a) { return a / 2 }
     function g (a) { return a + 2 }
     function h (a) { return a * 2 }
-    const i = compose(f, g, h)
-    expect(i(1)).toBe(2)
+    const i = pipe(f, g, h)
+    expect(i(1)).toBe(5)
   })
 })