+++ src/uncurried/eqBy.js
@@ -1,0 +1,3 @@
+export default function eqBy (f, a, b) {
+  return f(b) === f(a)
+}